针对遥控器逻辑做出修改，yaw轴pid基本上调好，双板通信频率改为500hz。现在的问题是云台跟随底盘模式没有效果，dm电机的offset有些问题需要解决。后续还想考虑通过双陀螺仪来提高底盘跟随精度。后续在想一下全麦底盘结算。

+++ chassis/application/robot_def.h
@@ -9,6 +9,7 @@
  *
  */
 #pragma once // 可以用#pragma once代替#ifndef ROBOT_DEF_H(header guard)
+#include <math.h>
 #ifndef ROBOT_DEF_H
 #define ROBOT_DEF_H
 
@@ -214,7 +215,7 @@ typedef struct
 typedef struct
 {
     attitude_t gimbal_imu_data;
-    uint16_t yaw_motor_single_round_angle;
+    float yaw_motor_single_round_angle;
 } Gimbal_Upload_Data_s;
 
 typedef struct
